header changes

+++ src/main/java/com/shvid/react/packedobject/PackedString.java
@@ -20,7 +20,7 @@ public final class PackedString implements PackedObject {
 	
 	public void format(byte[] blob) {
 		if (initCapacity > 0) {
-			int ptr = PackedFactory.allocate(blob, initCapacity + PackedData.getCapacitySize());
+			long ptr = PackedFactory.allocate(blob, initCapacity + PackedData.getCapacitySize());
 			PackedData.setCapacity(blob, ptr, initCapacity);
 			PackedData.setByte(blob, ptr + PackedData.getCapacitySize(), 0, initCapacity, (byte)0);
 			data.setPtr(blob, ptr);
@@ -32,7 +32,7 @@ public final class PackedString implements PackedObject {
 	
 	public void format(long address) {
 		if (initCapacity > 0) {
-			int ptr = PackedFactory.allocate(address, initCapacity + PackedData.getCapacitySize());
+			long ptr = PackedFactory.allocate(address, initCapacity + PackedData.getCapacitySize());
 			PackedData.setCapacity(address, ptr, initCapacity);
 			PackedData.setByte(address, ptr + PackedData.getCapacitySize(), 0, initCapacity, (byte)0);
 			data.setPtr(address, ptr);
